+++ packages/knip/src/compilers/mdx.ts
@@ -1,9 +1,8 @@
-import { fencedCodeBlockMatcher, importMatcher, importsWithinFrontmatter } from './compilers.js';
+import { fencedCodeBlockMatcher, importMatcher } from './compilers.js';
 import type { HasDependency } from './types.js';
 
 // https://mdxjs.com/packages/
 const mdxDependencies = [
-  'astro',
   '@mdx-js/esbuild',
   '@mdx-js/loader',
   '@mdx-js/mdx',
@@ -15,17 +14,8 @@ const mdxDependencies = [
   'remark-mdx',
 ];
 
-// Fields in frontmatter that could contain imports
-const frontmatterImportFields = ['layout'];
-
 const condition = (hasDependency: HasDependency) => mdxDependencies.some(hasDependency);
 
-const compiler = (text: string) => {
-  const imports = text.replace(fencedCodeBlockMatcher, '').matchAll(importMatcher);
-
-  const frontmatterImports = [importsWithinFrontmatter(text, frontmatterImportFields)];
-
-  return [...imports, ...frontmatterImports].join('\n');
-};
+const compiler = (text: string) => [...text.replace(fencedCodeBlockMatcher, '').matchAll(importMatcher)].join('\n');
 
 export default { condition, compiler };
